requirements.txt
nameko
attrs==19.3.0
cffi==1.13.2
Click==7.0
cryptography==2.8
importlib-metadata==1.3.0
more-itertools==8.0.2
mysql-connector-python==8.0.18
packaging==20.0
Pillow==7.0.0
pkg-resources==0.0.0
pluggy==0.13.1
protobuf==3.11.2
py==1.8.1
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.2
python-dotenv==0.10.3
python-dotenv-run==0.1.4
six==1.13.0
wcwidth==0.1.8
zipp==0.6.0
